Finished implementing multiple failures in one view change
Added a num_installed counter to the DerechoSST to decouple VIDs from
the number of committed changes that have been installed (applied to a
new view). Compiled and tested. Does not handle more than one join at a
time yet, due to the limited space in the SST for joiner_ip.