Implemented multiple joins in one view-change
This also entailed a fix to the logic of
tcp_connections::add_connection, which erroneously assumed that the
first node to connect to its listening socket would have the desired ID,
passed as other_id. (This assumption may be violated due to a race
condition if more than one new member attempts to connect at the same
time).