Reworked new-member initialization to account for subgroup state
The existing RPC code assumed that the group leader could initialize new
members with all the state of replicated RPC objects, but with subgroups
the group leader may not have all the object state because it may not be
a member of all subgroups. The leader now sends new members a vector of
node IDs that were "shard leaders" in the previous view, and each "shard
leader" is responsible for sending its subgroup's RPC object state.