Changed Message's "sender" field to refer to ID, not rank
This fixes a bug in the RPC message handler, which expects to get the
sender's ID but was actually getting its within-subgroup rank. Also
fixed a memory leak I noticed in SST: the raw pointer snapshot is never
deleted, and it refers to memory allocated in SST (copied from the SST
rows when a snapshot is made).