Fixed bugs in ragged edge cleanup introduced by subgroups
For some reason these weren't caught until now. The ragged edge cleanup
logic needed to change now that each subgroup has its own shard-local
leader that determines which messages are stable for that subgroup; this
was already reflected in the fact that global_min_ready is a vector,
rather than a single boolean, but the global_min_is_ready predicate did
not handle this and still assumed there was only a single leader.