Slight refactor of subgroup layout vector, bug fix
Created a new experiment testing out a smarter subgroup-allocation
function. This required refactoring the subgroup layout vector to own
the SubViews directly as values, rather than as unique_ptrs (so that the
vector could be copied).

Also fixed a bug in multicast_group in which wedge() attempted to clean
up predicate handles for subgroups that the current node is not a member
of; the predicate handles are now stored in contiguous lists independent
of subgroup numbers.