Fixed a bug related to sync_with_members in sst/multicast.h
When we are creating the SST only for the members of a subgroup, we
shouldn't be doing a sync with all the members. This will obviously
lead to an infinite wait. The particular instance in which it showed
itself was when we created 3 subgroups containing three members each
of a total of 10 members. The node which didn't belong to any of the
subgroups never called sync_with_members.