Changed epoch-termination logic to wait for an adequate view
The next view will not be installed until it will be adequately
provisioned according to the SubView allocation functions. Instead, the
group will stay wedged in the current view and wait for more joins to be
proposed by the leader, until there are enough new members for the next
view to be adequate. First step towards implementing the changes
required for issue #8.