Fixed bugs in persistence trigger and leader initialization
The persistence trigger in MulticastGroup neglected to check whether a
global persistence callback had actually been provided (even though it
defaults to nullptr in the user-facing code), so when I tried to test my
code with a simple non-persisted group this caused a crash unrelated to
my changes. Also, fixed a bug where I neglected to initialize
last_suspected in the leader, and refactored typed_subgroup_test to put
its definitions in a header file.