Fixed a bug in RPCManager, added some more testing
RPCManager and PendingResults had some race conditions relating to
fulfilling the "pending map" and "promises map." The local RPC handler
could attempt to handle a self-reply to an RPC message before the map
was fulfilled, and the P2P RPC thread could attempt to handle an
incoming reply to an RPC message before the map was fulfilled. Both
cases would cause an error (such as map out-of-bounds) in
PendingResults::set_value().