Fixed a bug: Constructing a temporary ExternalCaller had side effects
The constructor for RemoteInvoker gets a mutable reference to
RPCManager's receivers map, and uses operator[] to insert a new entry
into this map for its receive handler. However, since operator[]
overwrites any existing entry with the same key, this means a second
constructor call to RemoteInvoker for the same function will delete the
receivers entry for the earlier RemoteInvoker instance. This is a
problem because external_callers.emplace() constructs a temporary
ExternalCaller while testing to see if one already exists, and the
constructor for the temporary ExternalCaller replaces the recievers
entry for the existing, "real" ExternalCaller (which remains in the map
because emplace() does nothing if the map already contains an
ExternalCaller for the specified subgroup).