Integrated Weijia's changes with view-change code
Changed the code for new members joining an existing group to make them
first send the length of their existing persisted log (if any) and to
make the leader use this value to change how much of a persistent
object's log is serialized in to_bytes. Also fixed some minor errors in
Persistence and created a new test case.