Refactored identical definitions of struct Bytes into a header.
All of these experiments started with the exact same definition of a
class/struct called Bytes, and they all needed to be updated when the
ByteRepresentable interface changed. This header allows them to all
include the same definition without retyping the same code 6 times.

There are two other experiment files that use almost the same Bytes
object, except it has a copy constructor defined. If these experiments
don't depend on the lack of a copy constructor, maybe that definition
could be merged into this bytes_object.h.