Added try_connect and try_accept to TCP sockets
Added a try_connect function for tcp::socket that allows the caller to
specify a timeout, and also gives up immediately on an error instead of
continuing to retry forever. Added a try_accept function for
connection_listener that allows the caller to specify how long to wait
for an incoming connection before giving up. These will be helpful in
implementing failure recovery code (issue #8) since restarting nodes
will need to attempt connections to other nodes that may not have
restarted.