Removed "wait for adequately provisioned" code from experiments
Removed all the code blocks that repeatedly catch
subgroup_provisioning_exception until a sufficient number of members
have joined the group, because that's not necessary any more. The GMS
will not install an inadequate view, so the experiments will not proceed
until all the nodes have joined and the first view is adequately
provisioned.