Started integrating new "truncate log" feature into Replicated
In order for Weijia's truncate() function in Persistent<T> to be called
during restart, the Replicated<T>'s that contain Persistent<T> fields
need a new function that will call it on each field. I implemented this
in a similar way to the existing Replicated::trim() function that calls
trim() on each Persistent<T> field.