Refactored TCP connections and ReplicatedObject access. Closes #30.
The TCP connections pool now lives in Group and is shared with
RPCManager and ViewManager. ViewManager now gets a reference to the list
of ReplicatedObject references, which is still owned/managed by Group
(it turns out Group still needs to update this list when it adds or
removes a Replicated<T>, so ViewManager can't own it). This means the
send_object upcall can be implemented entirely inside ViewManager.

It turns out the receive_objects method still needs to reside within
Group, though, because it's used by both the new-view callback and the
non-leader Group constructor. Attempting to move this inside ViewManager
would probably result in code duplication, so I'm leaving it in Group
and keeping ViewManager's initialize_objects callback for now.