Disambiguated partial and total restarts, implemented changes for #28
Added some extra communication between the leader and a joining node.
The leader now informs the joining node if it is doing a total restart,
so that the joining node doesn't have to guess from the presence of a
logged View on disk (which would also be there if only that node had
failed). The leader also explicitly tells the joiner if it's trying to
rejoin with the same ID in a non-total-restart case, instead of waiting
for the view-change logic to crash. Finally, a non-leader that gets a
join request will use this same JoinResponse step to inform the client
that a different node is the leader.

Note, none of this is tested yet, so there are probably bugs.