Fixed bugs in the restart code, and changed how View is serialized
The standard mutils-serialization functions for View now serialize a lot
more of its fields, including the subgroup information maps, because
that information needs to be saved on disk in order for restart to work.
When sending the View to a new member, we now use StreamlinedView to
serialize only some of View's fields, knowing that the rest will be
reconstructed by the receiver.

Making View serialize subgroup_ids_by_type also required implementing
mutils-serialization methods for std::type_index, which is not a POD,
and refactoring the way different headers are included so that these
methods would be visible to mutils-serialization.