Implemented more restart logic based on Ken's e-mail
The restart leader now combines ragged trims and longest-log reports
from all the subgroups and shards (it has only logged a ragged trim for
the subgroups it belongs to), but requires at least one member from each
subgroup and shard from the last known view to contact it before it will
restart. Also, the restart leader now keeps track of which node in each
subgroup and shard has the longest log (based on reported ragged trims)
and designates that node as the "shard leader" for the purposes of
send_objects and receive_objects. This ensures that the nodes with the
longest logs send their persistent state to nodes with shorter logs
before anyone commits to this ragged trim and restart view.