Fixed a bug in p2p_connections.cpp
Closes #59. Although what Weijia reported wasn't a bug, as you are
supposed to start the nodes one by one and not simulteanously,
p2p_query_test also fails node 5 immediately and then resumes when it
rejoins as node 7. This failed because of a small bug where we should
be returning the node id instead of the rank. My bad.
Weijia and I will work on making this experiment more robust, so that
you can start all the nodes at the same time.