Stopped saving or attempting to fulfill PendingResults<void>. Fixes #65
Added the is_query paramter to finish_rpc_send as well as
finish_p2p_send, and made finish_rpc_send skip enqueuing the
PendingResults in toFulfillQueue if the message is a Send. This allowed
me to remove the logic from rpc_message_handler that detected and
removed PendingResults<void> from toFulfillQueue.