Fixed more bugs in total restart's handling of initial state transfer
The restart leader didn't attempt to receive logs for its replicated
objects, even though it might not have the latest state (unlike the case
for a normal leader, which never needs to receive state when it starts).
Also, await_rejoining_nodes moved the rejoining nodes' sockets to
proposed_join_sockets, even though they didn't need to be there and this
caused a subsequent crash in the next (real) view change.