Fixed an embarrassing bug in DefaultSubgroupAllocator
The operator() function created a previous_assignment vector before
attempting to assign nodes to all the subgroups, so if assignment failed
(due to not enough nodes to populate all the subgroups) the
previous_assignment vector would exist but not have all its entries. This
would cause the vector index operation to crash on the next invocation
of operator(), since it assumes that if previous_assignment exists it is
the right size.