Refactored restart leader logic, implemented more failure handling
Most of the code and state related to the restart leader's logic is now
stored in the RestartLeaderState class, and some logic that is shared
by both the restart leader and restart joiners is now in RestartState.
This also allowed me to more easily implement failure handling for cases
where a node crashes after connecting to the leader. The previous system
of checking the TCP connections explicitly before sending the View did
not work because TCP write/send calls can succeed even if the socket has
closed or failed.