Fixed bugs in the last commit, and a bug in TCP's Socket
The last commit turned out to have some errors that showed up right away
in testing. Also, the TCP Socket wrapper class had a long-standing bug:
if the remote end of a socket is closed unexpectedly, a write will
generate a SIGPIPE signal and crash the entire program instead of
returning an error code. Instead we need to use send() with the
MSG_NOSIGNAL parameter so that this type of error returns an error code
that we can handle.