Combined ordered_send and ordered_query, eliminated partial updates
The Replicated::ordered_send function is now the only way to send
totally-ordered RPC function calls to a shard, and it will return a
QueryResults<void> if the RPC function does not return anything.
QueryResults<void> has been implemented to allow the client to wait for
the RPC message to be delivered in the current View, which is the
functinality of QueryResults::get() for all other QueryResults objects.
Finally, the ordered_send function no longer allows the caller to
specify a subset of group members to send the RPC message to, since that
functionality is implemented by subgroups and shards, and each shard
should always get all updates for the state it owns. This fixes #4 and
fixes #75.