Fixed #61: RPC Opcodes no longer use std::type_index
I implemented the fix discussed in issue #61. RemoteInvocableClasses now
store a unique "type ID" integer that they use in each Opcode instead of
a std::type_index to identify the class that an RPC function belongs to.
This type ID is the index of the user-provided replicated object type
within the type paramters provided to derecho::Group, which works
because all RPC clients are within the same top-level Group and were
compiled with the same template parameters in the same order.