Added get_subgroup_members to Group interface
This new function is a corrollary to get_members for subgroups: it reads
the members list from one or more SubViews of the current View
(depending on whether the desired subgroup has one or several shards).
It provides a much more straightforward way for a node to determine
which subgroup it has been assigned to, other than repeatedly attempting
to call get_subgroup and catching the invalid_subgroup_exception.