one bugfix in ObjectStore.cpp and another in the design of persistence_manager.h
We used Group::replicated_objects in persistent_manager to find the Replicated<T> of a corresponding subgroup_id, which is invalid because in replicated_objects, the Replicated<T> is index by the subgroup_index (NOT subgroup_id). This bug didn't show up because in our previous tests on persistent feature, we have only one subgroup. The bugfix is move from Group::replicated_objects to Group::objects_by_subgroup_id. The latter is even faster without need to traverse object types.