Refactored View commits to use a 2-phase commit. Closes #106.
Now that state transfer can successfully happen before the first View is
committed, the leader needed to use a real 2-phase commit to commit the
restart View. (Formerly, the message sending the View counted as the
first phase of the commit, and we sent the second phase message as soon
as every node accepted the View, before doing state transfer).

State transfer before the View is installed is now tested and working.