Fixed bug in send_objects_to_new_members caused by changes in #106
I moved the call to this method up before next_view was installed, but
the method internally assumed that next_view had already been swapped
with curr_view. Now it explicitly takes as a parameter the View to use
as the "new view." (Also made a similar change to some other methods
that made assumptions about curr_view).