Fixed a bug I introduced back in my implementation of #71
When I changed subgroup allocation to run only on the leader, I somehow
left in the assumption that next_unassigned_rank should not be
serialized because it will be set on each receiving node when that node
runs the subgroup allocation function. In fact, next_unassigned_rank
must be serialized so that any node that takes over as leader has the
same information as the previous leader (namely, curr_view and
next_unassigned_rank) with which to compute the next view.

Note that adding next_unassigned_rank to the serialized fields of View
requires that mutils-serialization/SerializationMacros.hpp be
regenerated to include support for up to 13 fields. I don't know how we
want to handle this, but right now you'll have to run the following
command inside third_party/mutils-serialization:

./generate_macros.py 13