Another major refactor of the subgroup allocation function
My first attempt was inefficient because the function still had to
return a new subgroup_shard_layout_t on the second invocation, even if
that layout (a 2-d vector of SubViews) was unchanged from the first
invocation. It was also awkward to retrieve the existing allocation from
phase 1 in the function's second invocation, since it had to be
retrieved from curr_view using a subgroup ID that the function didn't
know. Now the subgroup allocation function is void, and places its
output in a map that is passed in as a parameter, which makes it easy to
edit that output in-place on the second invocation.