Minor fix: Added a necessary include to derecho_internal.h
All of our files that use the whenlog() macro implicitly depend on
including mutils/macro_utils.hpp, but the only one that actually
includes it is rpc_utils.h. This only works because we just happen to
include things in the right order such that rpc_utils.h is included
before anything that uses whenlog(). It would be much safer to include
macro_utils.hpp as a "global dependency" in derecho_internal.h, since,
along with the other definitions in that file, it must be present in
every Derecho source file.