Replaced whenlog macros with newer dbg_default_debug macros
Since we're using the default logger everywhere anyway, we should use
the macros in utils/logger.hpp that access it directly instead of
copying the pointer to those local "logger" variables.

Also, created "rls_default_*" macros that still execute in release mode,
for the rare cases where we want to log an error or print a message even
in release mode.