Partial fix for #16
This at least takes care of the most glaring problem described in #16,
which is that receiving an RPC message with an invalid Opcode causes the
recipient to crash when it's actually the sender's fault. Now the
recipient will simply ignore a messgae with an Opcode it doesn't know
how to handle, which allows it to keep running (though the sender may
still hang waiting for a reply that never comes).