Changed default subgroup allocator to not reassign nodes
It turns out that reassigning a node from one subgroup or shard to
another is not well supported by the rest of Derecho. The default
subgroup allocation function now ensures that nodes are never reassigned
to a different role, even if that means marking a view "inadequate"
unnecessarily (i.e. the view could become adequate by reassigning a node
from a shard with extra replicas to a shard with insufficient replicas).