Completed implementing option to disable partitioning safety
This, along with Weijia's commit, allows a user to disable Derecho's
logic for detecting a partitioning event by setting a Boolean option in
the config file. If the option DISABLE_PARTITIONING_SAFETY is set to
True, we will never throw an exception for a "potential partitioning
event." This means it will be possible for a 2-node group to go down to
1 node and continue running, but it also means that a real network
partition will allow the partitioned nodes to keep running and diverge
from the rest of the group. This fixes #112.