Fixed a major bug in report_failure
At some point, somebody incorrectly reused the variable "r" for both a
loop counter and the rank of the node who failed. This meant that the
SST put() call at the end of report_failure actually pushed an invalid
entry in suspected[], since at that point "r" had been incremented to
equal suspected.size() in the for loop.