Implemented changes for #92
Upon discovering that the old leader has failed, the new leader now
waits for every non-failed node to suspect the old leader before taking
any actions as the leader. The new leader also scans the SST for changes
proposed by previous leaders and re-proposes them before adding any of
its own proposals. The one thing left to do is to add the ability to
propose two separate successive views, which is not possible with the
current structure of the changes list.

I also made some minor code improvements, like using references to
gmsSST directly when possible instead of aliasing through curr_view, and
renaming the rank_of_leader() function to reflect the fact that it
recomputes the leader every time.