Reworked view change logic to fix #126
The group leader will now explicitly mark the last proposed change in a
View once it is reasonably certain it has processed all outstanding
leaves and joins in a batch; subsequent leaves and joins will be
proposed for the next view, which means that the group may transition to
an already-wedged view and then immediately start another view-change. A
new leader will re-propose the old leader's changes up until the
end-of-view marker, then propose its own changes in a new view. (This
also fixes #92).

Also, the assignment of nodes to subgroups and shards is now computed
only on the leader, instead of simultaneously by every member. All
non-leader nodes download the new View from the leader, instead of just
newly joining members. This makes it easier to enforce the leader's
batching of proposals, but will also make it easier to support more
complex subgroup-assignment functions in the future.

These changes are mostly tested, but may still contain some bugs because
further testing is blocked by unrelated bugs #144 and #141.