Reworked new-leader takeover logic to properly fix #92
After changing view proposals to use explicit batches (for issue #126),
the way a new leader copied proposals from prior leaders needed to
change too. The new leader will now only copy a prior leader's proposal
if that proposal includes an end-of-view marker, and any proposed
changes after the end-of-view marker will be rewritten to be proposals
from the new leader.