Fixed two related bugs in the total restart code
The total restart code did not properly account for the case where the
restart leader had not been a member of the last committed View. When
computing the restart view, it assumed that the restart leader was
already a member of curr_view (when actually the restart leader might
need to "join" in the restart view), and when initializing objects for
state transfer, it used curr_view rather than restart_view to determine
which subgroup objects to construct (which will fail if the restart
leader is not a member of any subgroup in curr_view). I fixed both of
these bugs so that total restart completes successfully if the restart
leader was not in the last curr_view.

Also, in a minor cosmetic change, I inlined the truncate_persistent_logs
function.