Replaced failure-handling macros with template functions (#151)
For the most part, these should be drop-in replacements. In a few cases,
I replaced FAIL_IF_NONZERO_RETRY_ON_EAGAIN with manual checking of the
return code, because the macro was actually overkill for that situation
(for example, in the post_remote_write methods, the post_remote_send
function called will never actually return EAGAIN).