Fixed another bug related to #153. Also added a more polished unit test
Restarting non-leader nodes attempted to truncate logs for subgroups
they didn't belong to, because truncate_logs incorrectly assumed that
curr_view->my_subgroups had been initialized and could be used to
determine whether this node belonged to a subgroup. (The my_subgroups
field of curr_view is not initialized until the view is committed and
initialize_multicast_groups is called).

Testing with the new unit test I wrote, log_restart_test (which is a
more polished version of total_restart_test), I can confirm that a group
with both persistent and non-persistent subgroups can successfully
restart using the total-restart process. This should be enough to close
issue #153.