Replaced source-built spdlog with packaged spdlog dependency
Ubuntu 19.04 and later contains a spdlog package that is new enough
(1.3.1) to use for our needs. Thus, there is no need to instruct users
to build it from source and install it in /usr/local/ (which is
generally fragile and error-prone); they should just use the system
package.

I made some minor updates to utils/logger to make it compile with spdlog
1.3.1, updated CMakeLists.txt to use the more standard invocation of
find_package(), and updated the README's instructions for installing
dependencies.