Replaced ports tuple with a ports struct
The std::tuple used to store IP addresses and ports for each group
member (in, for example, View) was getting very long and unwieldy now
that we have 5 different ports to keep track of. We all agreed it should
be replaced with a struct, so I did that (along with adding a few
comments and other cosmetic improvements).

This is a semantically-equivalent drop-in replacement, so no visible
functionality should change as a result; it just makes the code
prettier.