Added support for a "list of restart leaders" config option
In order to have a variable-size list of IP addresses in a single config
option, I had to write a string-splitting function that could be used to
split the resulting string around the commas. GetPot also has built-in
support for "vector options" that we could use, but that would require
changing the way the config options are stored in the Conf class (right
now each option can only be associated with a single string).