Refactored ViewManager startup to use restart leader in restart mode
This is the start of a major rewrite that will eventually support the
priority list of restart leaders, with automatic fallback when one of
them fails. For now, we only use the first element of the restart
leaders list and ignore the rest; it was hard enough to change
ViewManager and Group's startup code to differentiate the "restart
leader" from the "normal leader," which were assumed to be exactly the
same node before now.

Changes include:
 - ViewManager only has one constructor now, and figures out within it
   whether it should act as a leader or a joiner
 - ViewManager exposes a boolean function is_starting_leader() to help
   Group's constructor decide which code path to follow
 - The leader_connection socket is now a class member of ViewManager,
   stored as as unique_ptr, and no longer needs to be passed around to
   different functions as parameters. The downside is it must be
   explicitly deleted in finish_setup() in order to close the socket.