Fixed #141 and prevented deadlocks on node failure
Weijia determined that Infiniband completion events will have a
corrupted value in the "work request ID" field (wr_id or op_context) if
the remote node failed during a one-sided write, so both
verbs_poll_completion() and lf_poll_completion() should not attempt to
use this field if the completion is an error event. Instead, we just
drop the "bad" completion events and wait for the sender to time-out
waiting for them, which will still result in concluding that the remote
node failed.

After implementing this fix we discovered that detecting a node failure
by timeout could result in a deadlock because an in-progress send could
still be attempting to post a write to the failed node that will never
complete. (The sending thread will hold msg_state_mtx, which blocks the
predicate thread from completing a message-delivery trigger, which
prevents the view-change predicates from firing to remove the failed
node from the view). So we also added the ability to report a failure to
both Verbs and LibFabric-based sst::resources objects, which will cause
them to abort any in-progress remote operations rather than keep
retrying.

On an unrelated note, removed an obsolete .gitignore file from the tcp
directory.