Added checks for restart leader failure to ViewManager constructor
The phases of ViewManager setup that occur in the constructor (but not
those that occur in later methods) now check to see if the restart
leader has failed and gracefully recover by trying again at the next
restart leader on the list, rather than always throwing an exception any
time there's a socket error. A node that started restarting as a
non-leader can now become the restart leader if it detects the leader
has failed and then finds that it is the next "backup" restart leader.

I also refactored most of the constructor's code into two more helper
functions (fresh_initial_setup() and restart_initial_setup()) because
the constructor was getting too long and unwieldy.

At this point, restart leader failures that occur after the ViewManager
constructor has finished (e.g. during the "receive_objects" or
"prepare_initial_view" phases) will still result in a crash, because
it's impossible to go back to the beginning of the constructor and start
over once the constructor has finished. Handling these failures will
require more extensive rewriting of ViewManager and Group.