Added a config option to disable the backup restart leaders
At Ken's suggestion, the "fail-over" behavior during restart can be
disabled with a config option, and this option is set to "disabled" by
default. This will mostly return Derecho to its old behavior, except
that the single restart leader can still be set to a different node than
the normal-case leader (previously, these had to be the same node).