Refactored ViewManager's constructor to allow retrying after a failure
The ViewManager "constructor code" is now split into the actual
constructor, which just initializes member variables using its
parameters, and an "init" method, which does the initial setup of
contacting the leader and receiving the first view (or becoming the
leader and deciding the first view). The first_init method in turn calls
restart_to_initial_view if it discoveres a logged View on disk, which
contains the logic for determining which node to treat as the restart
leader. This allows Group's constructor to go back and run
restart_to_initial_view again, effectively re-running what used to be
the second half of ViewManager's constructor, if it discovers a leader
failure during total restart.

Unfortunately, fail-over to a backup restart leader during the "2-phase
commit" part of total restart still doesn't work. Even though the
failure is detected correctly during view_manager.check_view_committed,
the nodes have already done construct_objects and receive_objects in the
tentative restart view which has now been aborted. This means they
might not have instances of the same Replicated Objects that their
logged Views and RaggedTrims claim they do when they go back and run
restart_to_initial_view() again.

In order for this to actually work, there needs to be some way of
"tentatively" doing construct_objects while still keeping around the
logs of the previous View's objects in case of a failure. This seems
like it will be hard to implement.