Fixed bugs in failure handling during restart
Even without the "backup restart leaders" feature, it turns out failure
handling during the 2-phase-commit part of total restart never worked
correctly: If the leader detected a failure and sent an "abort" message,
the other nodes would not retry state transfer because construct_objects
would erroneously conclude that there was no need to download Replicated
Object state the second time it was called.

At this point restart with a single leader seems to work as advertised,
regardless of when non-leader failures occur.