Rewrote the TCP classes to use exceptions for error reporting
Rather than have methods that return false to indicate an error, a more
flexible way to report errors from TCP operations is to have them throw
exceptions specialized to indicate what type of error has occurred. This
also greatly simplifies blocks of code where we care about the success
or failure of a series of TCP operations - instead of checking the
return value of every method call, we can wrap the entire sequence of
operations in a try-catch block.

I created an exception hierarchy representing some common types of TCP
socket errors, rewrote the TCP socket functions to throw the correct
type of exception in cases where they used to return false, and rewrote
the parts of Derecho that use tcp::socket to catch these exceptions
rather than checking for a "success" return.