Minor cosmetic fix to ViewManager's tcp_sockets
The tcp_connections object being shared between ViewManager and Group
was a holdover from back before we used RDMA for P2P messages. Now that
RPCManager doesn't need any TCP connections, ViewManager is the only
class that uses these sockets, and they're only for state transfer. So I
moved tcp_sockets to be stored inside ViewManager, and gave it an
accessor method for the one time Group needs to use it (in
receive_objects).