Added a wrapper library for OpenSSL and a test program for it
In order to use cryptography such as digital signatures in the Derecho
code, we'll need to use a cryptography library, and OpenSSL is widely
available and a de facto standard. Plus, it seems like we already depend
on it for the persistence layer. Unfortunately, though, OpenSSL is a C
library that is awkward and cumbersome to use from C++. I wrote a
wrapper library that encapsulates the OpenSSL signature and message
digest functions within C++ classes that properly handle memory
management and error detection, similarly to how the "tcp" module wraps
the C-only TCP sockets library in C++-friendly classes and functions.
This library will reside in a new subfolder named "openssl," although to
avoid any issues with CMake detecting the real OpenSSL library, I named
the module "openssl_wrapper" in CMakeLists.txt.