Replaced persistence manager callbacks with direct method calls
The "persistence manager callbacks" were a convoluted way to let
MulticastGroup call the make_version and post_persistence_request
methods on PersistenceManager without a reference to the
PersistenceManager object itself. This was probably done back when
ViewManager was hard to construct, in order to avoid trying to pass a
reference to PersistenceManager into ViewManager. However, now that
ViewManager is much easier to construct, it turns out to be very
straightforward to just give ViewManager and MulticastGroup a reference
to PersistenceManager that is passed in at construction time. This
allows ViewManager and MulticastGroup to just call make_version() and
post_persistence_request() directly on PersistenceManager, instead of
going through a callback that wraps these functions in lambdas.

As a side effect of this change, I had to re-order a bunch of includes
and forward declarations to avoid a circular include dependency between
ViewManager and PersistenceManager, which now have references to each
other.