Preliminary implementation sketch for adding sign() to Persistent
After a discussion with Weijia, I think I figured out how to thread a
new "sign" function into the Persistent/PersistentRegistry call
hierarchy. This represents a preliminary implementation so that Weijia
has something to work with; it doesn't actually work yet, but it at
least compiles. A new sign() function has been added to Replicated,
Persistent, and PersistentRegistry, and the persist() function in all of
those places has been modified to accept as input the signature
generated by the sign() function.