Revised implementation sketch for adding sign() to Persistent
My earlier attempt was wrong because it didn't account for the fact that
persist() would attempt to batch multiple versions, each of which needs
its own signature. The new setup explicitly separates out the two-phase
signing process for each version, so that a signature for each version
can be added to the log before persist() is called. It also moves
ownership of the Signer to the Replicated object that calls into
PersistentRegistry, so that the process of getting the number of
versions in a batch, signing the batch, and then persisting the batch
can be encapsulated in the call to Replicated::persist().

This is still a work in progress and doesn't actually run yet.