rename Persistent::getLastPersisted() to Persistent::getLastPersistedVersion()
The previous name is a little confusing. This function does return the VERSION of last persisted log entry. Its comment was incorrect by saying the return value is an INDEX. Another related fix is, in the implementation of Persistent::register_callbacks(), the 4th function (get the latest persisted versions) now binding to getLastPersistedVersion() instead of getLatestVersion(). Please note that getLatestVersion() will return the same value as getLastPersistedVersion during system booting/recovery, which is the only time the function is called so far. That's why it did not caused any issue. To Edward: It's better to rerun the recovery tests on the changes.