Added chaining to signatures, and started implementing verification
When PersistentRegistry signs a new version, it now includes the
previous version's signature as part of the data that is signed. This
creates a "chain" of signatures, where each new version can only be
validated if the previous version also has a valid signature. I
implemented this with Weijia's help during a meeting.

Also started creating a framework for verifying signatures in a log,
with corresponding methods in Replicated, Persistent, and
PersistentRegistry. Unfortunately, this doesn't work yet because of the
signature chains: we need a method for getting the "previous version" in
a log (which is not necessarily version-1), and we need to figure out
how each node will retrieve another node's signature on the previous
version if it's going to verify the signature that node posted in the
SST for the current version.

Also added a new test file for the signed log that creates persistent
objects with multiple fields, since none of our previous tests used
objects with multiple fields but this is an important edge case for the
signed log.